% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterflyplot.R
\name{butterflyplot}
\alias{butterflyplot}
\alias{butterflyplot}
\alias{folded_butterflyplot}
\title{Butterfly plot}
\usage{
butterflyplot(x, y, col1 = "#FF000060", col2 = "#3200D360",
  border1 = "black", border2 = "black", breaks = 100)

folded_butterflyplot(x, y, col1 = "#FF000060", col2 = "#3200D360",
  border1 = "black", border2 = "black", breaks = 100)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric/factor/character vector with 2 levels.}

\item{col1, col2, border1, border2}{Character string color.}

\item{breaks}{Number of breaks.}
}
\description{
Generates a butterfly or folded butterfly plot
}
\examples{
butterflyplot(iris$Sepal.Length[1:100], droplevels(iris$Species[1:100]), breaks = 30)
folded_butterflyplot(iris$Sepal.Length[1:100], droplevels(iris$Species[1:100]), breaks = 30)
}
\keyword{butterflyplot,}
\keyword{folded_butterflyplot}
