% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\name{cd}
\alias{cd}
\title{Change directory}
\usage{
cd(dir)
}
\arguments{
\item{dir}{Unquoted text, usually a directory location}
}
\description{
Wrapper function intended to reduce typing.  Sends argument as a character string to \code{setwd} function.
}
\examples{
# Move up to parent directory
cd(..)

# Not run: Move to child directory
# cd(child)
}
\keyword{setwd}
