% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jby.R
\name{doj}
\alias{doj}
\alias{jby}
\alias{doj}
\title{Within data.table calculations among pipe chains}
\usage{
jby(data, expr_j, expr_by)

doj(data, expr_j)
}
\arguments{
\item{data}{A data.table object, say dt}

\item{expr_j}{The \code{j} expression for dt[,j,by]}

\item{expr_by}{The \code{by} expression for dt[,j,by]}
}
\description{
Perform a within data.table calculation dt[,j,by] or dt[,j] within a pipe chain (\%>\%) .
}
\details{
Allows dt[,j,by] or dt[,j] calculations within a pipe chain, such as

iris \%>\% as.data.table \%>\% doj(mean(Sepal.Length)) \%>\% print 
iris \%>\% as.data.table \%>\% jby(mean(Sepal.Length),Species) \%>\% print
}
\examples{
require(data.table); require(dplyr);
iris \%>\% as.data.table \%>\% jby(mean(Sepal.Length),Species) \%>\% print
require(data.table); require(dplyr);
iris \%>\% as.data.table \%>\% doj(mean(Sepal.Length)) \%>\% print
}
\keyword{doj}
\keyword{jby,}
