% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_consort.R
\name{draw_consort}
\alias{draw_consort}
\title{Draw Consort}
\usage{
draw_consort(data, arrows, phantom = NULL)
}
\arguments{
\item{data}{data.frame with the following columns}

\item{arrows}{data.frame with the following columns}

\item{phantom}{data.frame with the following columns}
}
\description{
Draws a flowchart or consort diagram
}
\examples{
#
}
\keyword{consort}
\keyword{flowchart,}
