% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_iguanatex.R
\name{engine_iguanatex}
\alias{engine_iguanatex}
\title{Provides iguanatex functionality for knitr}
\usage{
engine_iguanatex(options)
}
\arguments{
\item{options}{List.  Standard input of a knitr engine.}
}
\description{
Provides iguanatex functionality for knitr
}
\details{
Mimics the IguanaTex tool for PowerPoint.  Requires \code{pdflatex} and \code{dvisgm}.

Compiles Latex code into and svg image and inserts it to the document.
}
\keyword{knitr}
