% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit_rcs.R
\name{explicit_rcs}
\alias{explicit_rcs}
\title{Create a model formula with explicit knots}
\usage{
explicit_rcs(data, formula)
}
\arguments{
\item{data}{A data.frame / data.table object}

\item{formula}{A model formula which includes an \code{rcs} component, such as \code{y ~ age + rcs(bmi, 3)}}
}
\description{
Create a model formula with explicit knots
}
\details{
Returns a model formula with the number of knots replaced by explicit knot locations
}
\examples{
explicit_rcs(iris, Sepal.Length ~ rcs(Sepal.Width, 3))
}
\keyword{explicit_rcs}
