% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatp.R
\name{formatp}
\alias{formatp}
\title{Yet another p-value format function}
\usage{
formatp(x, digits = 3, sig = 0.05, sig_marker = c("", ""))
}
\arguments{
\item{x}{A p-value}

\item{digits}{The number of digits to display of the p-value.}

\item{sig}{The alpha or significance levels.  Used for formating.}

\item{sig_marker}{A length 2 character vector.  The formating for p-values < sig.  The first entry is placed before the p-value, the second entry is placed after the p-value.  For example, <b>0.034</b>.}
}
\description{
Formats p-values the way I like them
}
\examples{
formatp(runif(1))
formatp(0.0001)
formatp(0.023, sig_marker = c("☺ ", " ★ I'm addicted to p-values ★"))
formatp(0.051, sig_marker = c("☺ ", " ★ I'm addicted to p-values ★"))
}
\keyword{formatp}
