% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sas_readfile.R
\name{generate_sas_readfile}
\alias{generate_sas_readfile}
\alias{sasify}
\alias{shorten_no_vowels}
\title{Generates .csv and .sas files from a data.table}
\usage{
generate_sas_readfile(
  data,
  filename_stub,
  shortener = function(x) {
     substr(x, 1, 30)
 }
)

sasify(
  data,
  filename_stub,
  shortener = function(x) {
     substr(x, 1, 30)
 }
)

shorten_no_vowels(x)
}
\arguments{
\item{data}{A data.table object}

\item{filename_stub}{A character string which will be the filename of the generated files, eg, \code{filename_stub.csv} and \code{filename_stub.sas}.}

\item{shortener}{A function which shortens variable names if longer than 30 characters}
}
\description{
Generates .csv and .sas files from a data.table
}
\details{
The function generates 3 files:
\itemize{
\item filename_stub.csv
\item filename_stub.sas
\item filename_stub-variable-listing.csv
}

The variable listing provides details on variable names, type, length, and the original variable name (unabridged_name) if
\itemize{
\item the variable name was changed to comply with variable name character count restrictions, or
\item non alpha-numeric characters were removed from the variable name.
}

\code{shorten_no_vowels} is a custom shortener which removes vowels prior to selecting the first 30 characters
}
\examples{
iris_dt <- as.data.table(iris)
generate_sas_readfile(iris_dt, "iris")

# Long variable names
# Default shortener simply cuts the variable name at 30 characters
iris_dt$very_long_variable_name_that_exceeds_character_count_restrictions <- 1
generate_sas_readfile(iris_dt, "iris2")

# Custom shortener
shorten <- function(x){
  y <- gsub("a|e|i|o|u","",x)
  substr(y,1,30)
}
generate_sas_readfile(iris_dt, "iris3", shortener = shorten)
}
\keyword{SAS}
