% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{\%|\%}
\alias{\%|\%}
\title{glue}
\usage{
a \%|\% b
}
\arguments{
\item{a}{A character string}

\item{b}{A character string}
}
\description{
Glue character strings together.
}
\details{
Adapted from http://adv-r.had.co.nz/Functions.html
}
\examples{
"string A" \%|\% " string B"
"string A" \%|\% " string B" \%|\% "string C"
}
\keyword{paste0}
