% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlify_rownames.R
\name{htmlify_rownames}
\alias{htmlify_rownames}
\title{Converts hogwash rownames and dimnames to HTML friendly names}
\usage{
htmlify_rownames(x)
}
\arguments{
\item{x}{An array or data.frame}
}
\description{
Converts hogwash rownames and dimnames to HTML friendly names
}
\details{
Leading spaces are replaced with \code{&nbsp;} so that the table renders properly in HTML
}
\examples{
A <- array(1:6, dim = c(3,2))
rownames(A) <- c("A","  B","  C")
colnames(A) <- c("X1","X2")
htmlify_rownames(A)
}
\keyword{htmlify_rownames}
