% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jby.R
\name{jby}
\alias{jby}
\title{Calculate dt[,j,by] within pipes (\%>\%)}
\usage{
jby(data, expr1, expr2)
}
\arguments{
\item{data}{A data.table object, say dt}

\item{expr1}{The \code{j} expression for dt[,j,by]}

\item{expr2}{The \code{by} expression for dt[,j,by]}
}
\description{
Perform a within data.table calculation dt[,j,by] within a dplyr pipe construction.
}
\details{
Allows dt[,j,by] calculations withing a pipe chain, such as

iris \%>\% as.data.table \%>\% jby(mean(Sepal.Length),Species) \%>\% print
}
\examples{
require(data.table); require(dplyr);
iris \%>\% as.data.table \%>\% jby(mean(Sepal.Length),Species) \%>\% print
}
\keyword{jby}
