% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelmap.R
\name{levelmap}
\alias{levelmap}
\title{Levelmap}
\usage{
levelmap(x, map, type = "c")
}
\arguments{
\item{x}{A vector whose values will be mapped to new categories}

\item{map}{A data.frame/data.table where the first variable is the old values and the second column is the new categories}

\item{type}{Either "c" for character or "f" for factor.  The parameter determines the type of vector that is returned.}
}
\description{
Creates character/categorical variable by mapping old values to new categories
}
\details{
The function perpetuates the input label attribute to the output.
}
\examples{
set.seed(234)
xm <- data.frame(
    A = 0:1
  , B = c("The zero category","The one category")
  , stringsAsFactors = FALSE)
x <- rbinom(50, 1, .4)
levelmap(x, xm)
levelmap(x, xm, type = "f")
}
\keyword{levelmap}
