% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgrep.R
\name{lgrep}
\alias{lgrep}
\title{Logical grep}
\usage{
lgrep(x, pattern, ...)
}
\arguments{
\item{x}{A character vector}

\item{pattern}{A pattern expression for \code{grep}}

\item{...}{Additional options sent to \code{grep}}
}
\description{
This functions creates a logical vector to indicate which elements of \code{x} include the \code{pattern}.
}
\examples{
lgrep(LETTERS,"A|B|Z")
"A|B|Z" \%charin\% LETTERS
}
\keyword{grep}
