% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot_ci.R
\name{lineplot_ci}
\alias{lineplot_ci}
\title{Line plot with confidence interval}
\usage{
lineplot_ci(
  x,
  y,
  y_lb,
  y_ub,
  data = NULL,
  add = FALSE,
  line_col = "darkblue",
  ci_col = "gray80",
  type = "l",
  xlim = NULL,
  ylim = NULL,
  log = "",
  xlab = NULL,
  ylab = NULL,
  ann = par("ann"),
  axes = TRUE,
  frame.plot = axes,
  panel.first = grid(),
  panel.last = NULL,
  asp = NA,
  ...
)
}
\arguments{
\item{x, y, y_lb, y_ub}{Column name or column number for the respective \code{data}}

\item{data}{data.frame with plotting data}

\item{add}{TRUE/FALSE.  TRUE adds the plot the previous plot, FALSE generates a new plot.}

\item{line_col}{Color of line.  Default is "darkblue".}

\item{ci_col}{Color of confidence interval.  Default is "gray80".}

\item{...}{Other plotting parameters sent to plot.}
}
\description{
Draws lines withs with confidence intervals
}
\examples{
ed <- data.frame(
    x = -10:10
  , y = (-10:10)^2
  , y_lb = (-10:10)^2 - 10
  , y_ub = (-10:10)^2 + 10
  , stringsAsFactors = FALSE)
plotstyle(style = upright)
lineplot_ci(x = 1, y = 2, y_lb = 3, y_ub = 4, data = ed)
}
\keyword{lineplot_ci}
