% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwith.R
\name{lwith}
\alias{lwith}
\title{Commands repeated on each list element}
\usage{
lwith(l, ...)
}
\arguments{
\item{l}{A list.  Each element of the list is a data.frame.}

\item{...}{A set of commands to be executed}
}
\description{
Repeats a set of commands for each element of a list.  The commands are executed within the \code{with} command, so that variable names need not be quoted.
}
\examples{
iris \%>\% 
  split(.$Species) \%>\% 
  lwith({plot(Sepal.Length, Sepal.Width, main = Species[1])})
}
\seealso{
\code{\link{with}}
}
\keyword{lwith}
