% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_info.R
\name{missingness_info}
\alias{missingness_info}
\title{Compile missing data summary of dataset}
\usage{
missingness_info(data, upper_limit = 5, max_vars = 5)
}
\arguments{
\item{data}{A data.frame object for which a missing data report will be generated}

\item{upper_limit}{A number.  The right tail of the frequency distribution reported in the "by observation" summary is truncated to [upper_limit, Inf) and is displayed as "upper_limit +".}

\item{max_vars}{A number.  Limits the list of variables reported in the second data frame to the first max_vars most frequently missing variables.  If there are multiple variables with the same number of missing values, all such the variables will be reported. (This means more than max_vars variables can appear in the output)}
}
\description{
Generates two missing data summaries.  The first is "by row" which reports the distribution of missing data within each row.  The second is "by column" and reports the missing data percentage within each column.
}
\details{
The output is a list with the "by row" summary and the "by column" summary.
}
\examples{
tmp <- iris; tmp[1,1] <- NA
missingness_info(tmp)
}
\keyword{missingness_info}
