% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namify.R
\name{namify}
\alias{namify}
\title{Namify}
\usage{
namify(x)
}
\arguments{
\item{x}{A character vector or a data.frame}
}
\description{
Converts hogwash variable names to proper variable names.
}
\details{
The following changes are made to the text
\itemize{
\item Change . to _
\item Remove leading/trailing spaces, (, )
\item Change \code{\%} to pct
\item Change interior space to underscore
\item Change all characters to lower case
\item Remove punctuation
\item If text begins with number, prefix with n
}
}
\examples{
df_with_proper_variable_names <- namify(iris)
}
\keyword{names}
