% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point_ci.R
\name{plot_point_ci}
\alias{plot_point_ci}
\title{Coefficient Plot}
\usage{
plot_point_ci(point, lb, ub, labels, data, vertical_lines = NULL,
  col = NULL, xlim = NULL, ...)
}
\arguments{
\item{point}{Column name or column number for the point estimate variable in \code{data}}

\item{lb, ub}{Column name or column number for the lower/upper bound variable in \code{data}}

\item{labels}{Column name or column number for estimate labels in \code{data}}

\item{data}{A data.frame object with point, lower bound, and upper bound estimates}

\item{vertical_lines}{A numeric vector or scalar.  Draws a reference line at each value in \code{vertical_lines}.}

\item{col}{NULL or column name or column number.  If NULL, all estimates will be plotted black.  If a column name or column number is provided, estimates will be plotted according to the color provided in the corresponding column of \code{data}}

\item{xlim}{NULL or numeric couplet.  Providing a numeric couplet will set the xlim of the plot.}

\item{...}{Additional options set to \code{points} and \code{lines}}
}
\description{
Draws point estimates with confidence intervals
}
\examples{
ed <- data.frame(
    variable = c("Age", "Gender: Male")
  , beta = c("1","1.3")
  , lb = c(".5",".4")
  , ub = c("1.5","2.2")
  , col = c("darkred","navy")
  , stringsAsFactors = FALSE)
plot_point_ci(
  point = "beta"
  , lb = "lb"
  , ub = "ub"
  , labels = "variable"
  , data = ed
  , vertical_lines = 0
  , col = "col"
  , xlim = c(-1,3)
  , pch = 16
)
}
\keyword{plot_point_ci}
