% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.R
\name{plot_setup}
\alias{plot_setup}
\title{Setup a plot}
\usage{
plot_setup(
  d,
  f,
  xlim = NULL,
  ylim = NULL,
  log = "",
  asp = NA,
  xaxs = "r",
  yaxs = "r",
  ...
)
}
\arguments{
\item{d}{A data.frame.}

\item{f}{Optional. A formula y ~ x. The limits of the plot region is calculated from the range of y and x.}

\item{xlim}{Optional. A 2 element vector, the limits of the x axis.}

\item{ylim}{Optional. A 2 element vector, the limits of the y axis.}

\item{log, asp, xaxs, yaxs}{Optional. Options sent in the plot.window command}

\item{...}{Optional. Options sent to the par command}
}
\description{
Sets up a plot based on variables in a dataset.  The intended use is for chained/piped commands.  Returns the same dataset that was passed in.
}
\examples{
iris \%>\% 
  plot_setup(Sepal.Length~Sepal.Width) \%>\% 
  split(.$Species) \%>\% 
  lwith({
    l1 <- lm(Sepal.Length~Sepal.Width)
    lines(Sepal.Width, predict(l1))
  })
  
}
\keyword{plot_setup}
