% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotstyle.R
\name{plotstyle}
\alias{plotstyle}
\title{A collection of plotting parameters}
\usage{
plotstyle(data = NULL, style = NULL, ...)
}
\arguments{
\item{data}{NULL or data.frame / data.table.  See details.}

\item{style}{NULL or a non-quoted character string which indicates the plot style.  See details for list of plot styles.}
}
\description{
Sets several plotting parameters by keyword.
}
\details{
If a data object is provided, the exact same object is returned.  This allows one to set plotting parameters within a pipe chain.

Plotting styles include:

NULL (Default)
tight
upright
}
\examples{
require(dplyr)
data.frame(x = rnorm(100), y = rnorm(100)) \%>\% 
  plotstyle(tight) \%>\% 
  plot
  
data.frame(x = rnorm(100), y = rnorm(100)) \%>\% 
  plotstyle(upright) \%>\% 
  plot
}
\keyword{plotstyle}
