% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimented_scatter.R
\name{regimented_jitter}
\alias{regimented_jitter}
\title{Regimented Jitter}
\usage{
regimented_jitter(x, y, diameter = 1, ...)
}
\arguments{
\item{x}{A scalar number; the x-value in the plot}

\item{y}{A numeric vector; the values to be jittered regimentally}

\item{diameter}{A scalar number; controls the diameter of the circles in the plot}

\item{...}{Parameters which are sent to the \code{symbols} command}
}
\description{
Adds points to boxplots in a regimented way
}
\details{
Returns invisibly cbind(y, jitter_x, jitter_y) where y is the original data and jitter_x and jitter_y are the jittered centers in the plot.
}
\examples{
set.seed(2346)
y <- rnorm(123)
boxplot(y)
regimented_jitter(1, y, bg = "gray80")
boxplot(y)
regimented_jitter(1, y, diameter = 1.5, bg = "gray80")
}
\keyword{regimented_jitter}
