% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_box.R
\name{scatter_box}
\alias{scatter_box}
\title{Scatter box}
\usage{
scatter_box(x, ...)
}
\arguments{
\item{x}{for specifying data from which the boxplots are to be
    produced. Either a numeric vector, or a single list containing such
    vectors. Additional unnamed arguments specify further data
    as separate vectors (each corresponding to a component boxplot).
    \code{\link{NA}}s are allowed in the data.}

\item{...}{For the \code{formula} method, named arguments to be passed to
    the default method.

    For the default method, unnamed arguments are additional data
    vectors (unless \code{x} is a list when they are ignored), and named
    arguments are arguments and \link{graphical parameters} to be passed
    to \code{\link{bxp}} in addition to the ones given by argument
    \code{pars} (and override those in \code{pars}). Note that
    \code{bxp} may or may not make use of graphical parameters it is
    passed: see its documentation.
  }

\item{diameter}{A scalar which controls the size of the circles in the scatter plot}

\item{bg}{The fill color of the circles.}
}
\description{
Generates a boxplot with a regimented scatter plot
}
\details{
Returns, invisibly, the same output list as boxplot with an additional item named \code{jitter} which is the output from \code{regimented_scatter}.
}
\examples{
x <- rnorm(123)
grp <- rnorm(123) > 0
scatter_box(x~grp)
}
\keyword{scatter_box}
