% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_to_box.R, R/summaryM_to_df.R
\name{slice_to_box}
\alias{slice_to_box}
\alias{slice_to_box}
\title{Slice to box}
\usage{
slice_to_box(x)

slice_to_box(x)
}
\arguments{
\item{x}{A character vector.}

\item{tbl}{A summaryM table object.}

\item{html_space}{TRUE/FALSE.  If TRUE, function will replace white space characters with html white space character.  When TRUE, indenting of category labels will be preserved in HTML output.}

\item{...}{Additional parameters sent to \code{Hmisc::print.summaryM}}
}
\description{
Converts Q1/Q2/Q3 format of quantiles (as generated by \code{summaryM}) to Q2 [Q1, Q3] format.

Convert a summaryM table to data.frame.
}
\examples{
require(Hmisc)
require(knitr)
getHdata(pbc)
tbl1_formula <- bili + albumin + stage + protime + sex + age + spiders ~ drug
tbl1_formula \%>\% 
  summaryM(data=pbc) \%>\% 
  summaryM_to_df(html_space = FALSE, digits = 3) \%>\% 
  sapply(slice_to_box) \%>\% 
  kable
require(Hmisc)
getHdata(pbc)
tbl1_formula <- bili + albumin + stage + protime + sex + age + spiders ~ drug
tbl <- tbl1_formula \%>\% 
summaryM(data=pbc) \%>\% 
summaryM_to_df
}
\seealso{
summaryM_to_df

slice_to_box
}
\keyword{summaryM}
