% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryM_to_df.R
\name{summaryM_to_df}
\alias{summaryM_to_df}
\title{summaryM table to data.frame}
\usage{
summaryM_to_df(tbl, html_space = TRUE, ...)
}
\arguments{
\item{tbl}{A summaryM table object.}

\item{html_space}{TRUE/FALSE.  If TRUE, function will replace white space characters with html white space character.  When TRUE, indenting of category labels will be preserved in HTML output.}

\item{...}{Additional parameters sent to \code{Hmisc::print.summaryM}}
}
\description{
Convert a summaryM table to data.frame.
}
\examples{
require(Hmisc)
getHdata(pbc)
tbl1_formula <- bili + albumin + stage + protime + sex + age + spiders ~ drug
tbl <- tbl1_formula \%>\% 
summaryM(data=pbc) \%>\% 
summaryM_to_df
}
\seealso{
slice_to_box
}
\keyword{summaryM}
