% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_column_consort.R
\name{two_column_consort}
\alias{two_column_consort}
\title{Two column consort}
\usage{
two_column_consort(variable, top_box_text)
}
\arguments{
\item{variable}{A character vector where each element is in the following format: "\code{row col description of exclusion}", for example "3 2 BMI < 25".}

\item{top_box_text}{"A character string that provides the text for the top box of the consort diagram, for example "HCV Patient Registry"}
}
\description{
Creates a simple consort diagram.  Is a wrapper for functions that generate more complex consort diagrams.
}
\examples{
data <- iris \%>\% 
namify \%>\% 
mutate(consort_variable = ifelse(
  petal_width < .15,
  "2 2 Petal width < 0.15",
  ifelse(
    petal_length > 6,
    "3 2 Petal length > 6",
    ifelse(
      sepal_width >= 3 & sepal_width < 4,
      "4 2 Sepal Width ∊ [3, 4)",
      "5 1 Study population"
    )
  )
))

two_column_consort(
  variable = data[,"consort_variable"],
  top_box_text = "Iris dataset"
)
}
\keyword{consort}
