% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findvarname.R
\name{findvarname}
\alias{findvarname}
\title{findvarname}
\usage{
findvarname(var, data = NULL)
}
\arguments{
\item{var}{A character string regex search expression}

\item{data}{A data.frame object}
}
\description{
List variable names which match a search expression
}
\examples{
findvarname("length", iris)
fvn("Sepal",iris)
}
\keyword{grep}
