% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_box.R
\name{scatter_box}
\alias{scatter_box}
\title{Scatter box}
\usage{
scatter_box(x, ...)
}
\arguments{
\item{x}{A formula which will be sent to boxplot}

\item{...}{Parameters which will be sent to boxplot}

\item{diameter}{A scalar which controls the size of the circles in the scatter plot}

\item{bg}{The fill color of the circles.}
}
\description{
Generates a boxplot with a regimented scatter plot
}
\details{
Returns, invisibly, the same output list as boxplot with an additional item named \code{jitter} which is the output from \code{regimented_scatter}.
}
\examples{
x <- rnorm(123)
grp <- rnorm(123) > 0
scatter_box(x~grp)
scatter_box(x~grp, diameter = 1.5)
}
\keyword{scatter_box}
