% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_to_box.R
\name{slice_to_box}
\alias{slice_to_box}
\title{Slice to box}
\usage{
slice_to_box(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Converts Q1/Q2/Q3 format of quantiles (as generated by \code{summaryM}) to Q2 [Q1, Q3] format.
}
\examples{
require(Hmisc)
require(knitr)
getHdata(pbc)
tbl1_formula <- bili + albumin + stage + protime + sex + age + spiders ~ drug
tbl1_formula \%>\% 
  summaryM(data=pbc) \%>\% 
  summaryM_to_df(html_space = FALSE, digits = 3) \%>\% 
  sapply(slice_to_box) \%>\% 
  kable
}
\seealso{
summaryM_to_df
}
\keyword{summaryM}
